requirements.txt
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
phue==1.1
PyYAML==6.0
six==1.16.0
Werkzeug==2.1.1
